+++ vitest.config.js
@@ -28,9 +28,9 @@ export default defineConfig({
              */
             thresholds: {
                 lines: 95,
-                statements: 95,
-                functions: 90,
-                branches: 88,
+                statements: 93,
+                functions: 95,
+                branches: 85,
             },
         },
     },